Improve user interface functionality
Implemented a more dynamic and responsive user interface to both the analytics and new-users components. In addition, a new section was added to the new-users.component.html to display individual user details. Global CSS variables were utilized for uniform styling across the components. Also, added user image path to the data file for display purposes. Regular design improvements were made to the various components to enhance the usability and visual appeal of the app.